docs(changelog): add Test Lab decomposition entry under Unreleased
