+++ docs/changelog.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Refactor** — Test Lab page (`frontend/src/pages/TestLab.jsx`) trimmed by ~450 lines via extraction of presentational sub-components and pure helpers (audit §3.1). `ProjIcon`, `PipelinePanel` + `PIPELINE_STAGES`, `LiveLog`, and `QueueRow` now live under `frontend/src/components/test-lab/`; sessionStorage persistence helpers (`loadPersistedRun` / `persistRun` / `clearPersistedRun` / `STORAGE_KEY` / `LOG_CAP`) move to `frontend/src/utils/testLabPersistence.js`; attachment MIME guard + size caps (`isTextMime` / `ACCEPTED_EXTENSIONS` / `MAX_ATTACHMENT_SIZE` / `MAX_TOTAL_ATTACHMENT`) move to `frontend/src/utils/testLabAttachments.js`. `QueueTab` now imports `QueueRow`, `EmptyState`, and the `Clock` icon directly instead of receiving them via DI props. No behaviour change — pure decomposition.
+
 - **SEC / Perf** — Web fonts (Inter, JetBrains Mono) self-hosted via `@fontsource` instead of `@import url('https://fonts.googleapis.com/...')`. The Google Fonts CDN saw every visitor's IP + User-Agent on every page load, blocking GDPR-compliant EU deployments (CJEU C-101/01 / Munich VG 2022) and preventing CSP hardening of `style-src` / `font-src` beyond `'self' data:`. Also removes a ~120ms render-blocking external request on cold loads and makes the app work in air-gapped enterprise deployments. Imports added to `frontend/src/main.jsx`; bundle size impact is offset by eliminating the external stylesheet round-trip.
 
 - **Security** — Defence-in-depth XSS sanitization on AI-generated markdown rendered through `dangerouslySetInnerHTML` in `frontend/src/pages/AIChat.jsx`, `frontend/src/pages/ChatHistory.jsx`, and other consumers of `frontend/src/utils/markdown.js`. The hand-rolled parser was already XSS-safe by construction (every span is escaped via `escapeHtml()` before transforms; no `[text](url)` link rule exists to introduce attacker-controlled `href` / `src`), but the audit recommended defence-in-depth via `DOMPurify` with an explicit `ALLOWED_TAGS` / `ALLOWED_ATTR` allowlist. Catches future grammar expansions (e.g. adding link rendering) that might forget to guard `javascript:` URLs.
